# gzhttp: Remove a few unneeded allocs