```
Before:
Benchmark2kJitter-32    	   55479	     21371 ns/op	  95.83 MB/s	    3575 B/op	      20 allocs/op

After:
Benchmark2kJitter-32    	   54948	     21599 ns/op	  94.82 MB/s	    3451 B/op	      16 allocs/op
```

Don't put too much into the speed. GC is not included. Most of the allocs are in `httptest.ResponseRecorder`.