@greatroar 2023-03-07
Looks like an improvement, but I don't see how the copy of the jitter would be unavoidable. I've rebased #767 on this PR as a proof of concept. In that version, SetHeader just sets all the fields, which is restrictive, but sufficient to implement HTB.